MULTI APPROVING 1/23/17
We need to have a better way of populating the user list. right now, it isn't possible to reload the user list while navigating around the program. I'm not sure how the initial user list is generated. I think it's from the userFactory.registerCbs function call in the usersController, but i made that part of the userCtrl a callable variable and tried to execute it again and the userlist didn't reload. Other than that, the batch process is working reasonable well. I can easily user this code for batch deactivations as well. We need to sort out the refresh of the user list first though as I need to clear the checkbox data which isn't possible as things are set up now since we're not using the scope provider.